It's ugly but it works. Data is now logged to .csv
If you check the main function at the bottom of the Generation class you
should get an idea about how it works.

The Generation class is now a complete mess of logging stuff., battle
stuff and regeneration stuff which could probably be tidied up a lot. My
main effort was to make it work. Now that it works I should probably
tidy it up, perhaps separating out the logic into a few separate
classes:

Populations (for holding the bots)
Tournaments (for running tournaments and multi-round tournaments)
Regeneration (for breeding)
Logging (a single place to configure how logging should work)

I'm aware that there are logging frameworks available but never having
done logging before it feels simpler just to write a very simple logging
class that outputs data to a .csv in an appropriate format.